pyproject.toml
[tool.poetry]
name = "Book_RAG"
version = "0.1.0"
description = ""
authors = ["Kirill <kirillpozmogov017@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.3"
pre-commit = "^4.0.1"
black = "^24.10.0"
flake8 = "^7.1.1"
langchain = "^0.3.12"
sentence-transformers = "^3.3.1"
faiss-cpu = "^1.9.0.post1"
requests = "^2.32.3"
openai = "^1.57.4"
jupyter = "^1.1.1"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
langchain-community = "^0.3.12"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
